using for loop
Kwao has 155 salespersons. Draw a flowchart and write a C++ program to input for each salesperson, name, region of sale, and number of units of product sold and to determine and output for each salesperson, sales value and net commission earned.  There are two regions: U and S. Value of sales is the product of units sold and price per unit of the product. Price per unit sold is sh. 1200.  Sales commission is paid at the rate of 16% of sales values for sales in Region U but 17% for sales in Region S. For the part of sales commission of over sh. 25000, sales commission tax is 15% of sales commission. Otherwise, sales commission tax is 0. Net sales commission is sales commission less tax. The program should also accumulate totals of sales value according to region of sales and should output them per region when the program execution is ending.