using while loop
Vipi deals in three types of products (X, Y and Z) and has several customers. Draw a flowchart and write a C++ program to input for each purchase, customer name, type of product and the number of units of the product purchased.  Output should include salesperson name and the amount he or she should pay. Price per unit of Product X is sh. 400, Product Y is sh. 530 while that of Z is sh. 390.  A buyer buys only one type of the product. For a purchase value of over sh. 15000, a customer gets discount of 10% of sales value. Otherwise, discount is 5%. These discounts apply only to Products X and Z. There are no discounts for Product Y. Purchases of Product Y attract VAT of 18% of the net purchase value (i.e. purchase value after discount if any). There is no VAT for Products X and Z. VAT is added to the amount to be paid by the customer after discount if any.